removed hardcoded unused item id value
