+++ client/src/offsets.h
@@ -123,11 +123,3 @@ namespace ParamOffsets {
     const std::vector<uintptr_t> ItemLotParam2_Chr = { 0xA8, 0x50, 0xD8, 0x0 };
 #endif
 }
-
-const std::vector<int> unused_item_ids = {
-   3400000, 3401000, 21001100, 21001101, 21001102, 21001103, 21600000,
-   21610000, 21620000, 21630000, 21640000, 21650000, 21660000, 21670000,
-   21680000, 21690000, 21700000, 21710000, 26590000, 26750000, 26770000,
-   26800000, 26900000, 27521000, 60375000, 65240000, 65250000, 65260000,
-   65270000, 65280000, 65290000, 900008182, 900008183
-};
